seperated cleardb
